werk werk
Will be using sets to try to do rhyming on a phrase.

First grab a set of words that rhymes with each word
via the API, then take the intersection, and blah blah
blah. My hope is that this will make it simpler than
using a bunch of nested loops to match (and rhyme) each
word to every other word ("many to many").

Then for each "family" of rhymes I have to pick a color
to highlight. So for example, "still," "bill," and "frills,"
would all be red, and "stack," "wack," and "lacking," would
all be green or whatever.

Then I would add the highlight tag to each word and embed
that on the page.