Tracking only code that I write
For the moment I write in ES6 and then
transpile to ES5 because I sometimes work on a machine
that only is compatible with ES5.

None of the code in the js/ folder is written by me,
rather it is the output of Babel. All the code I write
is in src/, I only want GitHub to track what I write.